Add files via upload
This is an assembler program for a made up assembly language. it is ready to activate on ubuntu machinces, using the makefile. after that, just type in the bash "assembler file_name1 file_name2 ..." as many files as you need. the files shoud have a signeture of ".as" in the end of their name, but dont use it on running. I uploaded a file exmple of the assembly languge. have a good one!